Move retarget logic to its own file
